+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.0.6
 
-  * 
+  * Event.add\_column method gains a required attirbute (default false) to control whether parents are required to enter something
+  * Event::Column gains parent\_required attribute (default false)
 
 ## Version 1.0.5